# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Chronivaro REST API
  description: |
    REST API specification for Chronivaro, a Resource-Order-Activity time tracking and absence management system built on Strolch.
    Provides endpoints for time tracking, timers, presence overview, absence requests, period closing, reports, and administrative management.
  version: 0.1.0-SNAPSHOT
  contact:
    name: Chronivaro
    url: https://github.com/eitch/Chronivaro
servers:
  - url: /
    description: Current server instance
paths:
...
  /chronivaro/v1/absences:
    get:
      summary: List absences
      description: Retrieves absences with filtering by employee, date range, or state.
      operationId: getAbsences
      tags:
        - Absences
      security:
        - CookieAuth: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OffsetParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: employeeId
          in: query
          description: Filter by employee ID
          required: false
          schema:
            type: string
        - name: from
          in: query
          description: Start of period (ISO date or date-time)
          required: false
          schema:
            type: string
        - name: to
          in: query
          description: End of period (ISO date or date-time)
          required: false
          schema:
            type: string
        - name: state
          in: query
          description: Filter by absence state (REQUESTED, APPROVED, REJECTED, CANCELLED)
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Absences list retrieved successfully
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/X-Correlation-Id'
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/AbsenceDto'
                  - $ref: '#/components/schemas/PagedAbsenceDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: Request new absence
      description: Creates a new absence request.
      operationId: createAbsence
      tags:
        - Absences
      security:
        - CookieAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AbsenceCreateDto'
      responses:
        '200':
          description: Absence request submitted successfully
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Correlation-Id:
              $ref: '#/components/headers/X-Correlation-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AbsenceDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /chronivaro/v1/absences/{id}:
    get:
      summary: Get absence by ID
      description: Retrieves a single absence by identifier.
      operationId: getAbsenceById
      tags:
        - Absences
      security:
        - CookieAuth: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Absence identifier
          schema:
            type: string
      responses:
        '200':
          description: Absence retrieved successfully
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Correlation-Id:
              $ref: '#/components/headers/X-Correlation-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AbsenceDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      summary: Update absence
      description: Updates an existing absence request. Supports optimistic concurrency via `If-Match`.
      operationId: updateAbsence
      tags:
        - Absences
      security:
        - CookieAuth: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Absence identifier
          schema:
            type: string
        - $ref: '#/components/parameters/IfMatchParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AbsenceCreateDto'
      responses:
        '200':
          description: Absence updated successfully
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Correlation-Id:
              $ref: '#/components/headers/X-Correlation-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AbsenceDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Remove or cancel absence
      description: Removes or cancels an existing absence request. Supports optimistic concurrency via `If-Match`.
      operationId: removeAbsence
      tags:
        - Absences
      security:
        - CookieAuth: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Absence identifier
          schema:
            type: string
        - $ref: '#/components/parameters/IfMatchParam'
      responses:
        '200':
          description: Absence removed successfully
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/X-Correlation-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponseDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /chronivaro/v1/absences/{id}/approve:
    post:
      summary: Approve absence
      description: Approves a pending absence request (supervisor/admin action). Supports optimistic concurrency via `If-Match`.
      operationId: approveAbsence
      tags:
        - Absences
      security:
        - CookieAuth: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Absence identifier
          schema:
            type: string
        - $ref: '#/components/parameters/IfMatchParam'
      responses:
        '200':
          description: Absence approved successfully
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Correlation-Id:
              $ref: '#/components/headers/X-Correlation-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AbsenceDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /chronivaro/v1/absences/{id}/reject:
    post:
      summary: Reject absence
      description: Rejects a pending absence request with an optional comment. Supports optimistic concurrency via `If-Match`.
      operationId: rejectAbsence
      tags:
        - Absences
      security:
        - CookieAuth: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Absence identifier
          schema:
            type: string
        - $ref: '#/components/parameters/IfMatchParam'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActionCommentDto'
      responses:
        '200':
          description: Absence rejected successfully
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Correlation-Id:
              $ref: '#/components/headers/X-Correlation-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AbsenceDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /chronivaro/v1/absences/{id}/cancel:
    post:
      summary: Cancel absence
      description: Cancels an approved or pending absence request. Supports optimistic concurrency via `If-Match`.
      operationId: cancelAbsence
      tags:
        - Absences
      security:
        - CookieAuth: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Absence identifier
          schema:
            type: string
        - $ref: '#/components/parameters/IfMatchParam'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActionCommentDto'
      responses:
        '200':
          description: Absence cancelled successfully
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Correlation-Id:
              $ref: '#/components/headers/X-Correlation-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AbsenceDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ==========================================
  # Periods & Monthly Closing
  # ==========================================
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: strolch.authorization
      description: Strolch session cookie containing the user authentication token.
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer token containing the Strolch user certificate ID.
  headers:
    ETag:
      description: Entity version validator for optimistic concurrency control (e.g., W/"1" or "1").
      schema:
        type: string
        example: '"1"'
...
    X-Correlation-Id:
      description: Unique identifier for tracking and correlating requests across logs and client telemetry.
      schema:
        type: string
        format: uuid
        example: '4a1b0289-53e7-4b72-bdf4-f8b898ecb817'
  parameters:
    OffsetParam:
      name: offset
      in: query
      description: Zero-based item offset for pagination (defaults to 0).
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
        example: 0
    LimitParam:
      name: limit
      in: query
      description: Number of items to return (defaults to 50, maximum 1000).
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 50
        example: 50
    IfMatchParam:
      name: If-Match
      in: header
      required: false
      description: Target entity version for optimistic concurrency control.
      schema:
        type: string
        example: '"1"'
  responses:
    BadRequest:
      description: Bad request - Validation failure or malformed payload
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/X-Correlation-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDto'
    Unauthorized:
      description: Unauthorized - Authentication is missing or invalid
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/X-Correlation-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDto'
    Forbidden:
      description: Forbidden - Insufficient user privileges or role permissions
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/X-Correlation-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDto'
    NotFound:
      description: Not found - Requested resource does not exist
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/X-Correlation-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDto'
    Conflict:
      description: Conflict - Concurrency mismatch (stale If-Match) or business state violation
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/X-Correlation-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDto'
    InternalServerError:
      description: Internal server error - Unexpected system failure
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/X-Correlation-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDto'
  schemas:
...
    ErrorDto:
      type: object
      required:
        - error
        - code
        - message
        - correlationId
      properties:
        error:
          type: boolean
          example: true
        code:
          type: string
          example: 'VALIDATION_ERROR'
        message:
          type: string
          example: 'Invalid input parameters'
        correlationId:
          type: string
          example: '4a1b0289-53e7-4b72-bdf4-f8b898ecb817'
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldErrorDto'
    FieldErrorDto:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
          example: 'email'
        message:
          type: string
          example: 'Email address is invalid'
    SuccessResponseDto:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: 'Operation completed successfully'
    ActionCommentDto:
      type: object
      properties:
        comment:
          type: string
          example: 'Approved based on team coverage agreement'
...
    AbsenceDto:
      type: object
      properties:
        id:
          type: string
          example: 'abs-001'
        employeeId:
          type: string
          example: 'emp-001'
        employeeName:
          type: string
          nullable: true
          example: 'John Doe'
        personalNumber:
          type: string
          nullable: true
          example: 'PER-1001'
        teamName:
          type: string
          nullable: true
          example: 'Engineering'
        absenceTypeCode:
          type: string
          example: 'VACATION'
        absenceTypeName:
          type: string
          nullable: true
          example: 'Vacation'
        start:
          type: string
          format: date-time
          example: '2026-08-20T00:00:00+02:00'
        end:
          type: string
          format: date-time
          example: '2026-08-22T23:59:59+02:00'
        durationType:
          type: string
          enum: [DAYS, HALF_DAY, HOURS]
          example: 'DAYS'
        dayPart:
          type: string
          enum: [MORNING, AFTERNOON]
          nullable: true
          example: 'MORNING'
        minutes:
          type: integer
          nullable: true
          example: 1440
        comment:
          type: string
          nullable: true
          example: 'Summer vacation'
        state:
          type: string
          enum: [REQUESTED, APPROVED, REJECTED, CANCELLED]
          example: 'APPROVED'
    AbsenceCreateDto:
      type: object
      required:
        - employeeId
        - absenceTypeCode
        - start
        - end
        - durationType
      properties:
        employeeId:
          type: string
          example: 'emp-001'
        absenceTypeCode:
          type: string
          example: 'VACATION'
        start:
          type: string
          format: date-time
          example: '2026-08-20T00:00:00+02:00'
        end:
          type: string
          format: date-time
          example: '2026-08-22T23:59:59+02:00'
        durationType:
          type: string
          enum: [DAYS, HALF_DAY, HOURS]
          example: 'DAYS'
        dayPart:
          type: string
          enum: [MORNING, AFTERNOON]
          example: 'MORNING'
        minutes:
          type: integer
          example: 1440
        comment:
          type: string
          example: 'Annual vacation'
    PagedAbsenceDto:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AbsenceDto'
        offset:
          type: integer
          example: 0
        limit:
          type: integer
          example: 50
        total:
          type: integer
          example: 15
        size:
          type: integer
          example: 15

    # ------------------------------------------
    # Periods
    # ------------------------------------------
...